Refs #1549. Updated implementation.

+++ test/blackbox/types/HelloWorldType.cpp
@@ -37,7 +37,7 @@ bool HelloWorldType::serialize(void* data, SerializedPayload_t* payload)
     payload->encapsulation = ser.endianness() == eprosima::fastcdr::Cdr::BIG_ENDIANNESS ? CDR_BE : CDR_LE;
 	//serialize the object:
 	hw->serialize(ser);
-	payload->length = (uint16_t)ser.getSerializedDataLength();
+	payload->length = (uint32_t)ser.getSerializedDataLength();
 	return true;
 }
